+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFilterTableSearchModel.java
@@ -171,10 +171,12 @@ final class DefaultFilterTableSearchModel<C> implements FilterTableSearchModel {
 			return;
 		}
 		Predicate<String> predicate = searchPredicate.get();
+		List<FilterTableColumn<C>> visibleColumns = columnModel.visible();
 		for (int row = 0; row < tableModel.getRowCount(); row++) {
-			for (int column = 0; column < tableModel.getColumnCount(); column++) {
-				if (predicate.test(tableModel.getStringAt(row, columnModel.getColumn(column).identifier()))) {
-					searchResults.add(new DefaultRowColumn(row, column));
+			for (int columnIndex = 0; columnIndex < visibleColumns.size(); columnIndex++) {
+				FilterTableColumn<C> column = visibleColumns.get(columnIndex);
+				if (predicate.test(tableModel.getStringAt(row, column.identifier()))) {
+					searchResults.add(new DefaultRowColumn(row, columnIndex));
 				}
 			}
 		}